changed memo alwell to be long

+++ Binomial.java
@@ -23,7 +23,7 @@ public class Binomial {
 	
 	// Computes the Binomial function, efficiently
 	public static long binomial(int n, int k) {
-		int[][] memo = new int[n + 1][k + 1];
+		long[][] memo = new long[n + 1][k + 1];
 
         
         for (int i = 0; i <= n; i++) {
@@ -35,7 +35,7 @@ public class Binomial {
         return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
